Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic>=2.0,<2.6
openai>=1.0,<2.0
python-dotenv==1.0.0
aiofiles==23.2.1
chromadb>=0.4.0 